refactor blog post listView into components
