+++ www/components/Navbar.tsx
@@ -32,7 +32,7 @@ const Navbar = () => {
     return (
       <li
         key={`navbar-link-to-${link.name}`}
-        className={`list-none p-2 text-center border-b-2 font-semibold ${
+        className={`list-none p-2 text-center border-b-2 text-xl ${
           isActivePath(currentPath, link.route)
             ? ` border-gray-200`
             : `border-gray-800`
@@ -46,9 +46,7 @@ const Navbar = () => {
   return (
     <div className="bg-gray-800">
       <div className="px-4 mx-auto lg:max-w-3xl text-gray-200 font-sourcecodepro flex flex-row py-2 items-center">
-        <h1 className="text-xl font-extrabold whitespace-nowrap ">
-          Ben Gordon
-        </h1>
+        <h1 className="text-xl font-extrabold whitespace-nowrap">Ben Gordon</h1>
         <div className="w-full flex flex-row justify-end">
           <nav className="max-w-3xl w-full flex flex-row justify-end">
             {navbarLinkJSX}
