+++ www/tailwind.config.js
@@ -16,6 +16,12 @@ module.exports = {
         sourcecodepro: ["Source Code Pro", "monospace"],
       },
     },
+    dropShadow: {
+      glow: [
+        "0 0px 20px rgba(255,255, 255, 0.35)",
+        "0 0px 65px rgba(255, 255,255, 0.2)",
+      ],
+    },
   },
   plugins: [],
   purge: ["./pages/**/*.{js,ts,jsx,tsx}", "./components/**/*.{js,ts,jsx,tsx}"],
